src/metadata.c: fix format for __u64 in metadata.c
On ppc64le, __u64 is typedef'd as 'unsigned long' rather than
'unsigned long long', so cast to unsigned long long fix the warning.

Fixes: https://github.com/linux-blktests/blktests/issues/233

+++ src/metadata.c
@@ -246,7 +246,7 @@ static int check_metadata(void *p, int intervals, int ref)
 
 				if (v != ref) {
 					fprintf(stderr, "reftag interval:%d expected:%x got:%llx\n",
-						i, ref, v);
+						i, ref, (unsigned long long)v);
 					return -1;
 				}
 				remaining -= sizeof(*tuple);
